// e2e/packages/e2e/src/types.ts
export type NexusSetUpOptions = {
  mock?: boolean;
  userName: string;
  passwd: string;
  seed?: string;
};

export type AddNetworkOpt = {
  name: string;
  url: string;
};
export type PopupPageHelper = {
  getNewPage: () => Promise<Page>;
  queryWhitelist: () => Promise<string[]>;
  removeWhitelistBySearch: (search: string) => void;
  queryNetworkList: () => Promise<string[]>;
  addNetwork: (addNetworkOpt: AddNetworkOpt) => void;
  removeNetworkByName: (name: string) => void;
  changeNetworkByName: (name: string) => void;
  queryConnected: () => Promise<boolean>;
  queryNickName: () => Promise<string>;
};

export type WalletManagerHelper = {
  getNewPage: () => Promise<Page>;
  importWallet: (page: Page, userName: string, mnemonic: string, password: string) => void;
  createANewWallet: (page: Page, userName: string, password: string) => Promise<string>;
};
export type NexusWallet = {
  popup: PopupPageHelper;
  approve: (passwd: string) => void;
  connect: () => void;
  cancel: () => void;
  walletManager: WalletManagerHelper;
  getNotificationPage: () => Promise<Page>;
  close: () => void;
};